backend/requirements.txt
Flask==3.1.1
google_api_python_client==2.176.0
google_auth_oauthlib==1.2.2
python-dotenv==1.1.1
Requests==2.32.4
gunicorn
pytz